Demo of codec switching: Fix changeType racing previous stream's init segment fetch
This change fixes the potential race that was resulting in parse error
(changeType(stream2) + append init segment (stream1)), in this case
making the changeType's parameter match the type of the init segment
subsequently appended.

ABR switch can occur during the asynchronous fetch of the previously
chosen stream's init segment. ABR switch modifies the mediaState in the
interim, so the changeType parameter used for appending the previously
chosen stream's init segment must be calculated before the init
segment's fetch yields control back to the browser.

This change also includes the originalId in the logPrefix_ to assist
further debugging.